mos-core: V0.6 Updates
Major functional additions and changes:
  - A new tool (mosview) is introduced.  This tool can set
    or query the mOS view of a process.
  - A new yod option (--mosview) is added to immediately set
    the mos view of the LWK process.
  - An improved NUMA reservation algorithm in yod which is
    both cleaner and more capable of supporting very large
    and complex node topologies.
  - yod and lwkctl are updated to properly support the
    newly added mOS views.
  - Improvements in the test plugins ability to capture
    request information and therefore improve detail in
    validation.
  - Improvements in the test plugin's ability to simulate
    larger and more varied topologies.
  - yod no longer affinitizes processes that are launched using
    the test plugin.  This helps with simulating topologies that
    are grossly different from the host.
  - Optimize mos_cpuset_is_empty() which has a measurable effect
    in launch times for large node topologies.

Minor changes and fixes:
  - Fixes to the RPM install script.
  - The previously deprecated K-scheduler code has been removed.
  - Fixed assorted typos in the man pages.
  - Cleanup and consistent usage of size_t in yod.
  - Improvements in yod tracing code.
  - Annotate fall-through legs of switch statements so that
    more recent compilers do not complain.

+++ mOS/tools/yod/layout.c
@@ -178,7 +178,7 @@ void _map_system_layout(void)
 {
 	struct map_type_t *nodes, *tiles, *cores;
 	mos_cpuset_t *tmp;
-	int i, N, t, T, c, C;
+	size_t i, N, t, T, c, C;
 
 	nodes = yod_get_map(YOD_NODE);
 	tiles = yod_get_map(YOD_TILE);
@@ -253,7 +253,7 @@ void _map_system_layout(void)
 			}
 		}
 
-		YOD_LOG(YOD_GORY, "SYSMAP: %s -> (%d,%d,%d)",
+		YOD_LOG(YOD_GORY, "SYSMAP: %s -> (%zd,%zd,%zd)",
 			mos_cpuset_to_list_validate(core), N, T, C);
 
 		assert(_system_map[COORD(N, T, C)] == NULL);
